Move helper functions into utils.py
