+++ hknweb/candidate/views.py
@@ -6,38 +6,15 @@ from django.core.exceptions import PermissionDenied
 from django.template.loader import render_to_string
 from django.contrib import messages
 from django.contrib.auth.models import User
-from django.contrib.auth.decorators import login_required, permission_required
-from django.utils.decorators import method_decorator
 from django.utils import timezone
 from django.conf import settings
-from django.contrib.staticfiles.finders import find
 from django.db.models import Q
 from dal import autocomplete
-from functools import wraps
-from random import randint
 
+from hknweb.utils import login_and_permission, method_login_and_permission, get_rand_photo
 from .models import OffChallenge, BitByteActivity, Announcement, CandidateForm
 from ..events.models import Event, Rsvp
 from .forms import ChallengeRequestForm, ChallengeConfirmationForm, BitByteRequestForm
-
-# decorators
-
-def login_and_permission(permission_name):
-    """ First requires log in, but if you're already logged in but don't have permission,
-        displays more info. """
-    def decorator(func):
-        return wraps(func)( # preserves function attributes to the decorated function
-                login_required(login_url='/accounts/login/')(
-                    # raises 403 error which invokes our custom 403.html
-                    permission_required(permission_name, login_url='/accounts/login/', raise_exception=True)(
-                        func # decorates function with both login_required and permission_required
-                    )
-                )
-            )
-    return decorator
-
-def method_login_and_permission(permission_name):
-    return method_decorator(login_and_permission(permission_name), name='dispatch')
 
 # views
 
@@ -274,18 +251,6 @@ class UserAutocomplete(autocomplete.Select2QuerySetView):
         return qs
 
 # HELPERS
-
-def get_all_photos():
-    """ This function is not used; it can be used to view all photos available. """
-    with open(find("candidate/animal_photo_urls.txt")) as f:
-        urls = f.readlines()
-    return [url.strip() + "?w=400" for url in urls]
-
-# images from pexels.com
-def get_rand_photo(width=400):
-    with open(find("candidate/animal_photo_urls.txt")) as f:
-        urls = f.readlines()
-    return urls[randint(0, len(urls) - 1)].strip() + "?w=" + str(width)
 
 def send_cand_confirm_email(request, challenge, confirmed):
     subject = '[HKN] Your officer challenge was reviewed'
